\name{huge-internal}
\title{Internal huge functions}
\alias{huge.mb}
\alias{huge.ct}
\alias{huge.glasso}
\description{Internal huge functions}
\author{
Tuo Zhao, Han Liu, Kathryn Roeder, John Lafferty, and Larry Wasserman \cr
Maintainers: Tuo Zhao<Tuo Zhao<tzhao5@jhu.edu>>
}
\details{
These are not intended for use by users. Please refer to \code{huge()}
}
