\name{plot.GECT}
\alias{plot.GECT}
\title{
Plot function for S3 class "GECT"
}
\description{
Plot sparsity level information for the graph path by GECT.
}
\usage{
\method{plot}{GECT}(x, ...)
}
\arguments{
  \item{x}{
An object with S3 class \code{"GECT"}
}
  \item{\dots}{
System reserved (No specific usage)
}
}
\author{
Tuo Zhao, Han Liu, Kathryn Roeder, John Lafferty, and Larry Wasserman \cr
Maintainers: Tuo Zhao<tourzhao@andrew.cmu.edu>; Han Liu <hanliu@cs.jhu.edu>
}
\seealso{
\code{\link{huge.GECT}}
}