\name{plot.glassoM}
\alias{plot.glassoM}
\title{
Plot function for S3 class "glassoM"
}
\description{
Plot sparsity level information for the graph path.
}
\usage{
\method{plot}{glassoM}(x, ...)
}
\arguments{
  \item{x}{
An object with S3 class \code{"glassoM"} 
}
  \item{\dots}{
System reserved (No specific usage)
}
}
\author{
Tuo Zhao, Han Liu, Kathryn Roeder, John Lafferty, and Larry Wasserman \cr
Maintainers: Tuo Zhao<tourzhao@gmail.com>; Han Liu <hanliu@cs.jhu.edu>
}
\seealso{
\code{\link{huge.glassoM}} and \code{\link{huge}}
}