\name{plot.huge}
\alias{plot.huge}
\title{
Plot function for S3 class "huge"
}
\description{
Plot sparsity level information and 3 typical sparse graphs from the graph path
}
\usage{
\method{plot}{huge}(x, align = FALSE, ...)
}
\arguments{
  \item{x}{
An object with S3 class \code{"huge"} 
}
  \item{align}{
If \code{align = FALSE}, 3 plotted graphs are aligned
}
  \item{\dots}{
System reserved (No specific usage)
}
}
\author{
Tuo Zhao, Han Liu, Kathryn Roeder, John Lafferty, and Larry Wasserman \cr
Maintainers: Tuo Zhao<Tuo Zhao<tzhao5@jhu.edu>>; Han Liu <hanliu@cs.jhu.edu>
}
\seealso{
\code{\link{huge}}
}