\name{plot.MBGEL}
\alias{plot.MBGEL}
\title{
Plot function for S3 class "MBGEL"
}
\description{
Plot sparsity level information for the graph path.
}
\usage{
\method{plot}{MBGEL}(x, ...)
}
\arguments{
  \item{x}{
An object with S3 class \code{"MBGEL"} 
}
  \item{\dots}{
System reserved (No specific usage)
}
}
\author{
Tuo Zhao, Han Liu, Kathryn Roeder, John Lafferty, and Larry Wasserman \cr
Maintainers: Tuo Zhao<tourzhao@gmail.com>; Han Liu <hanliu@cs.jhu.edu>
}
\seealso{
\code{\link{huge.MBGEL}} and \code{\link{huge}}
}