\name{plot.npn}
\alias{plot.npn}
\title{
Plot function for S3 class "npn"
}
\description{
Plot the empirical covariance matrices before and after the nonparanormal transformation
}
\usage{
\method{plot}{npn}(x,...)
}
\arguments{
  \item{x}{
An object with S3 class \code{"npn"} 
}
  \item{\dots}{
System reserved (No specific usage)
}
}
\author{
Tuo Zhao, Han Liu, Kathryn Roeder, John Lafferty, and Larry Wasserman \cr
Maintainers: Tuo Zhao<tourzhao@andrew.cmu.edu>; Han Liu <hanliu@cs.jhu.edu>
}
\seealso{
\code{\link{huge.npn}} and \code{\link{huge}}
}