\name{plot.roc}
\alias{plot.roc}
\title{
Plot function for S3 class "roc"
}
\description{
Plot the ROC curve for an object with S3 class \code{"roc"}
}
\usage{
\method{plot}{roc}(x, ...)
}
\arguments{
  \item{x}{
An object with S3 class \code{"roc"} 
}
  \item{\dots}{
System reserved (No specific usage)
}
}
\author{
Tuo Zhao, Han Liu, Kathryn Roeder, John Lafferty, and Larry Wasserman \cr
Maintainers: Tuo Zhao<tourzhao@gmail.com>; Han Liu <hanliu@cs.jhu.edu>
}
\seealso{
\code{\link{huge.roc}}
}