\name{plot.scr}
\alias{plot.scr}
\title{
Plot function for S3 class "scr"
}
\description{
Plot sparsity level information for the solution path by GECA.
}
\usage{
\method{plot}{scr}(x, ...)
}
\arguments{
  \item{x}{
An object with S3 class \code{"scr"} with \code{x$approx = TRUE}
}
  \item{\dots}{
System reserved (No specific usage)
}
}
\author{
Tuo Zhao, Han Liu, Kathryn Roeder, John Lafferty, and Larry Wasserman \cr
Maintainers: Tuo Zhao<tourzhao@gmail.com>; Han Liu <hanliu@cs.jhu.edu>
}
\seealso{
\code{\link{huge.scr}}
}