\name{plot.select}
\alias{plot.select}
\title{
Plot function for S3 class "select"
}
\description{
Plot the optimal graph by model selection
}
\usage{
\method{plot}{select}(x, ...)
}
\arguments{
  \item{x}{
An object with S3 class \code{"select"} 
}
  \item{\dots}{
System reserved (No specific usage)
}
}
\author{
Tuo Zhao, Han Liu, Kathryn Roeder, John Lafferty, and Larry Wasserman \cr
Maintainers: Tuo Zhao<tourzhao@andrew.cmu.edu>; Han Liu <hanliu@cs.jhu.edu>
}
\seealso{
\code{\link{huge.select}}
}