% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huge.generator.R
\name{plot.sim}
\alias{plot.sim}
\title{Plot function for S3 class "sim"}
\usage{
\method{plot}{sim}(x, ...)
}
\arguments{
\item{x}{An object with S3 class \code{"sim"}}

\item{\dots}{System reserved (No specific usage)}
}
\description{
Visualize the covariance matrix, the empirical correlation matrix, the adjacency matrix and the graph pattern of the true graph structure.
}
\seealso{
\code{\link{huge.generator}} and \code{\link{huge}}
}
