\name{plot.subgraph}
\alias{plot.subgraph}
\title{
Plot function for S3 class "subgraph"
}
\description{
Plot sparsity level information for the solution path by GEL.
}
\usage{
\method{plot}{subgraph}(x, ...)
}
\arguments{
  \item{x}{
An object with S3 class \code{"subgraph"} 
}
  \item{\dots}{
System reserved (No specific usage)
}
}
\author{
Tuo Zhao, Han Liu, Kathryn Roeder, John Lafferty, and Larry Wasserman \cr
Maintainers: Tuo Zhao<tourzhao@gmail.com>; Han Liu <hanliu@cs.jhu.edu>
}
\seealso{
\code{\link{huge.subgraph}}
}