\name{print.GECT}
\alias{print.GECT}
\title{
Print function for S3 class "GECT"
}
\description{
Print the information about the graph path information and length for GECT
}
\usage{
\method{print}{GECT}(x, ...)
}
\arguments{
  \item{x}{
An object with S3 class \code{"GECT"}
}
  \item{\dots}{
System reserved (No specific usage)
}
}
\author{
Tuo Zhao, Han Liu, Kathryn Roeder, John Lafferty, and Larry Wasserman \cr
Maintainers: Tuo Zhao<tourzhao@gmail.com>; Han Liu <hanliu@cs.jhu.edu>
}
\seealso{
\code{\link{huge.GECT}} and \code{\link{huge}}
}