\name{print.glassoM}
\alias{print.glassoM}
\title{
Print function for S3 class "glassoM"
}
\description{
Print information about the model usage and the solution path length
}
\usage{
\method{print}{glassoM}(x, ...)
}
\arguments{
  \item{x}{
An object with S3 class \code{"glassoM"}
}
  \item{\dots}{
System reserved (No specific usage)
}
}
\author{
Tuo Zhao, Han Liu, Kathryn Roeder, John Lafferty, and Larry Wasserman \cr
Maintainers: Tuo Zhao<tourzhao@andrew.cmu.edu>; Han Liu <hanliu@cs.jhu.edu>
}
\seealso{
\code{\link{huge.glassoM}} and \code{\link{huge}}
}