\name{print.huge}
\alias{print.huge}
\title{
Print function for S3 class "huge"
}
\description{
Print the information about the model usage, the graph path length, graph dimension, sparsity level
}
\usage{
\method{print}{huge}(x, ...)
}
\arguments{
  \item{x}{
An object with S3 class \code{"huge"}
}
  \item{\dots}{
System reserved (No specific usage)
}
}
\author{
Tuo Zhao, Han Liu, Kathryn Roeder, John Lafferty, and Larry Wasserman \cr
Maintainers: Tuo Zhao<tourzhao@gmail.com>; Han Liu <hanliu@cs.jhu.edu>
}
\seealso{
\code{\link{huge}} and \code{\link{huge}}
}