\name{print.MBGEL}
\alias{print.MBGEL}
\title{
Print function for S3 class "MBGEL"
}
\description{
Print information about the model usage and the graph path length
}
\usage{
\method{print}{MBGEL}(x, ...)
}
\arguments{
  \item{x}{
An object with S3 class \code{"MBGEL"}
}
  \item{\dots}{
System reserved (No specific usage)
}
}
\author{
Tuo Zhao, Han Liu, Kathryn Roeder, John Lafferty, and Larry Wasserman \cr
Maintainers: Tuo Zhao<tourzhao@gmail.com>; Han Liu <hanliu@cs.jhu.edu>
}
\seealso{
\code{\link{huge.MBGEL}} and \code{\link{huge}}
}