\name{print.npn}
\alias{print.npn}
\title{
Print function for S3 class "npn"
}
\description{
Print the information about the transformation function}
\usage{
\method{print}{npn}(x, ...)
}
\arguments{
  \item{x}{
An object with S3 class \code{"npn"}
}
  \item{\dots}{
System reserved (No specific usage)
}
}
\author{
Tuo Zhao, Han Liu, Kathryn Roeder, John Lafferty, and Larry Wasserman \cr
Maintainers: Tuo Zhao<tourzhao@andrew.cmu.edu>; Han Liu <hanliu@cs.jhu.edu>
}
\seealso{
\code{\link{huge.npn}} and \code{\link{huge}}
}