\name{print.roc}
\alias{print.roc}
\title{
Print function for S3 class "roc"
}
\description{
Print the information about true positive rates, false positive rates, the area under curve and maximum F1 score
}
\usage{
\method{print}{roc}(x, ...)
}
\arguments{
  \item{x}{
An object with S3 class \code{"roc"}
}
  \item{\dots}{
System reserved (No specific usage)
}
}
\author{
Tuo Zhao, Han Liu, Kathryn Roeder, John Lafferty, and Larry Wasserman \cr
Maintainers: Tuo Zhao<tourzhao@gmail.com>; Han Liu <hanliu@cs.jhu.edu>
}
\seealso{
\code{\link{huge.roc}}
}