\name{print.scr}
\alias{print.scr}
\title{
Print function for S3 class "scr"
}
\description{
Print the information about the solution path information and length for GECA or the preselected neighborhood size for the graph screening
}
\usage{
\method{print}{scr}(x, ...)
}
\arguments{
  \item{x}{
An object with S3 class \code{"scr"}
}
  \item{\dots}{
System reserved (No specific usage)
}
}
\author{
Tuo Zhao, Han Liu, Kathryn Roeder, John Lafferty, and Larry Wasserman \cr
Maintainers: Tuo Zhao<tourzhao@gmail.com>; Han Liu <hanliu@cs.jhu.edu>
}
\seealso{
\code{\link{huge.scr}}
}