\name{print.select}
\alias{print.select}
\title{
Print function for S3 class "select"
}
\description{
Print the information about the model usage, graph dimension, model selection criterion, sparsity level of the optimal graph
}
\usage{
\method{print}{select}(x, ...)
}
\arguments{
  \item{x}{
An object with S3 class \code{"select"}
}
  \item{\dots}{
System reserved (No specific usage)
}
}
\author{
Tuo Zhao, Han Liu, Kathryn Roeder, John Lafferty, and Larry Wasserman \cr
Maintainers: Tuo Zhao<Tuo Zhao<tzhao5@jhu.edu>>; Han Liu <hanliu@cs.jhu.edu>
}
\seealso{
\code{\link{huge.select}} and \code{\link{huge}}
}