\name{print.sim}
\alias{print.sim}
\title{
Print function for S3 class "sim"
}
\description{
Print the information about the sample size, the dimension and the pattern of the true graph strcture.
}
\usage{
\method{print}{sim}(x, ...)
}
\arguments{
  \item{x}{
An object with S3 class \code{"sim"}
}
  \item{\dots}{
System reserved (No specific usage)
}
}
\author{
Tuo Zhao, Han Liu, Kathryn Roeder, John Lafferty, and Larry Wasserman \cr
Maintainers: Tuo Zhao<tourzhao@gmail.com>; Han Liu <hanliu@cs.jhu.edu>
}
\seealso{
\code{\link{huge.generator}}
}