\name{summary.GECT}
\alias{summary.GECT}
\title{
Summary function for S3 class "GECT"
}
\description{
Summary the information about the graph path information and length for GECT
}
\usage{
\method{summary}{GECT}(object, ...)
}
\arguments{
  \item{object}{
An object with S3 class \code{"GECT"}
}
  \item{\dots}{
System reserved (No specific usage)
}
}
\author{
Tuo Zhao, Han Liu, Kathryn Roeder, John Lafferty, and Larry Wasserman \cr
Maintainers: Tuo Zhao<tourzhao@andrew.cmu.edu>; Han Liu <hanliu@cs.jhu.edu>
}
\seealso{
\code{\link{huge.GECT}} and \code{\link{huge}}
}