\name{summary.huge}
\alias{summary.huge}
\title{
Summary function for S3 class "huge"
}
\description{
Summary the information about the model usage, the solution path length, subgraph dimension, fullgraph dimension, sparsity level
}
\usage{
\method{summary}{huge}(object, ...)
}
\arguments{
  \item{object}{
An object with S3 class \code{"huge"}
}
  \item{\dots}{
System reserved (No specific usage)
}
}
\author{
Tuo Zhao, Han Liu, Kathryn Roeder, John Lafferty, and Larry Wasserman \cr
Maintainers: Tuo Zhao<tourzhao@andrew.cmu.edu>; Han Liu <hanliu@cs.jhu.edu>
}
\seealso{
\code{\link{huge}}
}