\name{summary.MBGEL}
\alias{summary.MBGEL}
\title{
Summary function for S3 class "MBGEL"
}
\description{
Summary information about the model usage and the graph path length
}
\usage{
\method{summary}{MBGEL}(object, ...)
}
\arguments{
  \item{object}{
An object with S3 class \code{"MBGEL"}
}
  \item{\dots}{
System reserved (No specific usage)
}
}
\author{
Tuo Zhao, Han Liu, Kathryn Roeder, John Lafferty, and Larry Wasserman \cr
Maintainers: Tuo Zhao<tourzhao@andrew.cmu.edu>; Han Liu <hanliu@cs.jhu.edu>
}
\seealso{
\code{\link{huge.MBGEL}} and \code{\link{huge}}
}