\name{summary.roc}
\alias{summary.roc}
\title{
Summary function for S3 class "roc"
}
\description{
Summary the information about true positive rates, false positive rates, the area under curve and maximum F1 score
}
\usage{
\method{summary}{roc}(object, ...)
}
\arguments{
  \item{object}{
An object with S3 class \code{"roc"}
}
  \item{\dots}{
System reserved (No specific usage)
}
}
\author{
Tuo Zhao, Han Liu, Kathryn Roeder, John Lafferty, and Larry Wasserman \cr
Maintainers: Tuo Zhao<tourzhao@andrew.cmu.edu>; Han Liu <hanliu@cs.jhu.edu>
}
\seealso{
\code{\link{huge.roc}} and \code{\link{huge}}
}