\name{summary.scr}
\alias{summary.scr}
\title{
Summary function for S3 class "scr"
}
\description{
Summary the information about the solution path information and length for GECA or the preselected neighborhood size for the graph screening
}
\usage{
\method{summary}{scr}(object, ...)
}
\arguments{
  \item{object}{
An object with S3 class \code{"scr"}
}
  \item{\dots}{
System reserved (No specific usage)
}
}
\author{
Tuo Zhao, Han Liu, Kathryn Roeder, John Lafferty, and Larry Wasserman \cr
Maintainers: Tuo Zhao<tourzhao@andrew.cmu.edu>; Han Liu <hanliu@cs.jhu.edu>
}
\seealso{
\code{\link{huge.scr}} and \code{\link{huge}}
}