\name{summary.stars}
\alias{summary.stars}
\title{
Summary function for S3 class "stars"
}
\description{
Summary the information about the solution path length and the degree of freedom's along the solution path.
}
\usage{
\method{summary}{stars}(object, ...)
}
\arguments{
  \item{object}{
An object with S3 class \code{"stars"}
}
  \item{\dots}{
System reserved (No specific usage)
}
}
\author{
Tuo Zhao, Han Liu, Kathryn Roeder, John Lafferty, and Larry Wasserman \cr
Maintainers: Tuo Zhao<tourzhao@gmail.com>; Han Liu <hanliu@cs.jhu.edu>
}
\seealso{
\code{\link{lasso.stars}}
}