\name{summary.subgraph}
\alias{summary.subgraph}
\title{
Summary function for S3 class "subgraph"
}
\description{
Summary information about the model usage and the solution path length
}
\usage{
\method{summary}{subgraph}(object, ...)
}
\arguments{
  \item{object}{
An object with S3 class \code{"subgraph"}
}
  \item{\dots}{
System reserved (No specific usage)
}
}
\author{
Tuo Zhao, Han Liu, Kathryn Roeder, John Lafferty, and Larry Wasserman \cr
Maintainers: Tuo Zhao<tourzhao@andrew.cmu.edu>; Han Liu <hanliu@cs.jhu.edu>
}
\seealso{
\code{\link{huge.subgraph}} and \code{\link{huge}}
}