\name{huge-internal}
\title{Internal huge functions}
\alias{huge.mbgel}
\alias{huge.gect}
\alias{huge.glasso}
\description{Internal huge functions}
\author{
Tuo Zhao, Han Liu, Kathryn Roeder, John Lafferty, and Larry Wasserman \cr
Maintainers: Tuo Zhao<tourzhao@gmail.com>; Han Liu <hanliu@cs.jhu.edu>
}
\details{
These are not intended for use by users. Please refer to \code{huge()}
}
