% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huge-package.R
\docType{package}
\name{huge-package}
\alias{huge-package}
\alias{_PACKAGE}
\title{High-Dimensional Undirected Graph Estimation}
\description{
A package for high-dimensional undirected graph estimation
}
\details{
\tabular{ll}{
  Package: \tab huge\cr
  Type: \tab Package\cr
  Version: \tab 1.2.7\cr
  Date: \tab 2015-09-14\cr
  License: \tab GPL-2\cr
  LazyLoad: \tab yes\cr
}
The package "huge" provides 8 main functions:\cr
(1) the data generator creates random samples from multivariate normal distributions with different graph structures. Please refer to \code{\link{huge.generator}}.\cr
(2) the nonparanormal (npn) transformation helps relax the normality assumption. Please refer to \code{\link{huge.npn}}.\cr
(3) The correlation thresholding graph estimation. Please refer to \code{\link{huge}}.\cr
(4) The Meinshausen-Buhlmann graph estimation. Please refer to \code{\link{huge}}.\cr
(5) The graphical Lasso algorithm using lossless screening rule. Please refer and \code{\link{huge}}.\cr
\cr
**Both (4) and (5) can be furether accelerated by the lossy screening rule preselecting the neighborhood of each node via thresholding sample correlation.
\cr
(6) The model selection using the stability approach to regularization selection. Please refer to \code{\link{huge.select}}.\cr
(7) The model selection using the rotation information criterion. Please refer to \code{\link{huge.select}}.\cr
(8) The model selection using the extended Bayesian information criterion. Please refer to \code{\link{huge.select}}.\cr
}
\references{
1.  T. Zhao and H. Liu. The huge Package for High-dimensional Undirected Graph Estimation in R. \emph{Journal of Machine Learning Research}, 2012\cr
2.  H. Liu, F. Han, M. Yuan, J. Lafferty and L. Wasserman. High Dimensional Semiparametric Gaussian Copula Graphical Models. \emph{Annals of Statistics},2012 \cr
3.  D. Witten and J. Friedman. New insights and faster computations for the graphical lasso. \emph{Journal of Computational and Graphical Statistics}, to appear, 2011.
4.  Han Liu, Kathryn Roeder and Larry Wasserman. Stability Approach to Regularization Selection (StARS) for High Dimensional Graphical Models. \emph{Advances in Neural Information Processing Systems}, 2010.\cr
5.  R. Foygel and M. Drton. Extended bayesian information criteria for gaussian graphical models. \emph{Advances in Neural Information Processing Systems}, 2010.\cr
6.  H. Liu, J. Lafferty and L. Wasserman. The Nonparanormal: Semiparametric Estimation of High Dimensional Undirected Graphs. \emph{Journal of Machine Learning Research}, 2009 \cr
7.  J. Fan and J. Lv. Sure independence screening for ultra-high dimensional feature space (with discussion). \emph{Journal of Royal Statistical Society B}, 2008.\cr
8.  O. Banerjee, L. E. Ghaoui, A. d'Aspremont: Model Selection Through Sparse Maximum Likelihood Estimation for Multivariate Gaussian or Binary Data. \emph{Journal of Machine Learning Research}, 2008.\cr
9.  J. Friedman, T. Hastie and R. Tibshirani. Regularization Paths for Generalized Linear Models via Coordinate Descent. \emph{Journal of Statistical Software}, 2008. \cr
10. J. Friedman, T. Hastie and R. Tibshirani. Sparse inverse covariance estimation with the lasso, \emph{Biostatistics}, 2007.\cr
11. N. Meinshausen and P. Buhlmann. High-dimensional Graphs and Variable Selection with the Lasso. \emph{The Annals of Statistics}, 2006.\cr
}
\seealso{
\code{\link{huge.generator}}, \code{\link{huge.npn}}, \code{\link{huge}}, \code{\link{huge.plot}} and \code{\link{huge.roc}}
}
\author{
Tuo Zhao, Han Liu, Haoming Jiang, Kathryn Roeder, John Lafferty, and Larry Wasserman \cr
Maintainers: Haoming Jiang<hjiang98@gatech.edu>;
}
