% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huge.ct.R
\name{huge.ct}
\alias{huge.ct}
\title{Graph estimation via correlation thresholding (ct)}
\usage{
huge.ct(x, nlambda = NULL, lambda.min.ratio = NULL, lambda = NULL,
  verbose = TRUE)
}
\description{
See more details in \code{\link{huge}}
}
\seealso{
\code{\link{huge}}, and \code{\link{huge-package}}.
}
