% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huge.mb.R
\name{huge.mb}
\alias{huge.mb}
\title{Meinshausen & Buhlmann graph estimation}
\usage{
huge.mb(x, lambda = NULL, nlambda = NULL, lambda.min.ratio = NULL,
  scr = NULL, scr.num = NULL, idx.mat = NULL, sym = "or",
  verbose = TRUE)
}
\description{
See more details in \code{\link{huge}}
}
\seealso{
\code{\link{huge}}, and \code{\link{huge-package}}.
}
