% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huge.tiger.R
\name{huge.tiger}
\alias{huge.tiger}
\title{Tuning-insensitive graph estimation}
\usage{
huge.tiger(x, lambda = NULL, nlambda = NULL, lambda.min.ratio = NULL,
  sym = "or", verbose = TRUE)
}
\description{
See more details in \code{\link{huge}}
}
\seealso{
\code{\link{huge}}, and \code{\link{huge-package}}.
}
