% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huge.R
\name{plot.huge}
\alias{plot.huge}
\title{Plot function for S3 class "huge"}
\usage{
\method{plot}{huge}(x, align = FALSE, ...)
}
\arguments{
\item{x}{An object with S3 class \code{"huge"}}

\item{align}{If \code{align = FALSE}, 3 plotted graphs are aligned}

\item{\dots}{System reserved (No specific usage)}
}
\description{
Plot sparsity level information and 3 typical sparse graphs from the graph path.
}
\seealso{
\code{\link{huge}}
}
