\name{plot.NPN}
\alias{plot.NPN}
\title{
Plot function for S3 class "NPN"
}
\description{
Plot the empirical covariance matrices before and after the nonparanormal transformation
}
\usage{
\method{plot}{NPN}(x,...)
}
\arguments{
  \item{x}{
An object with S3 class \code{"NPN"} 
}
  \item{\dots}{
System reserved (No specific usage)
}
}
\author{
Tuo Zhao, Han Liu, Kathryn Roeder, John Lafferty, and Larry Wasserman \cr
Maintainers: Tuo Zhao<tourzhao@gmail.com>; Han Liu <hanliu@cs.jhu.edu>
}
\seealso{
\code{\link{huge.NPN}} and \code{\link{huge}}
}