% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huge.roc.R
\name{plot.roc}
\alias{plot.roc}
\title{Plot function for S3 class "roc"}
\usage{
\method{plot}{roc}(x, ...)
}
\arguments{
\item{x}{An object with S3 class \code{"roc"}}

\item{\dots}{System reserved (No specific usage)}
}
\description{
Plot the ROC curve for an object with S3 class \code{"roc"}.
}
\seealso{
\code{\link{huge.roc}}
}
