% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huge.select.R
\name{plot.select}
\alias{plot.select}
\title{Plot function for S3 class "select"}
\usage{
\method{plot}{select}(x, ...)
}
\arguments{
\item{x}{An object with S3 class \code{"select"}}

\item{\dots}{System reserved (No specific usage)}
}
\description{
Plot the optimal graph by model selection.
}
\seealso{
\code{\link{huge.select}}
}
