\name{plot.sim}
\alias{plot.sim}
\title{
Plot function for S3 class "sim"
}
\description{
Visualize the covariance matrix, the empirical covariance matrix, the adjacency matrix and the graph pattern of the true graph structure
}
\usage{
\method{plot}{sim}(x, ...)
}
\arguments{
  \item{x}{
An object with S3 class \code{"sim"} 
}
  \item{\dots}{
System reserved (No specific usage)
}
}
\author{
Tuo Zhao, Han Liu, Kathryn Roeder, John Lafferty, and Larry Wasserman \cr
Maintainers: Tuo Zhao<tourzhao@gmail.com>; Han Liu <hanliu@cs.jhu.edu>
}
\seealso{
\code{\link{huge.generator}} and \code{\link{huge}}
}