% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huge.R
\name{print.huge}
\alias{print.huge}
\title{Print function for S3 class "huge"}
\usage{
\method{print}{huge}(x, ...)
}
\arguments{
\item{x}{An object with S3 class \code{"huge"}.}

\item{\dots}{System reserved (No specific usage)}
}
\description{
Print the information about the model usage, the graph path length, graph dimension, sparsity level.
}
\seealso{
\code{\link{huge}}
}
