\name{print.NPN}
\alias{print.NPN}
\title{
Print function for S3 class "NPN"
}
\description{
Print the information about the transformation function}
\usage{
\method{print}{NPN}(x, ...)
}
\arguments{
  \item{x}{
An object with S3 class \code{"NPN"}
}
  \item{\dots}{
System reserved (No specific usage)
}
}
\author{
Tuo Zhao, Han Liu, Kathryn Roeder, John Lafferty, and Larry Wasserman \cr
Maintainers: Tuo Zhao<tourzhao@andrew.cmu.edu>; Han Liu <hanliu@cs.jhu.edu>
}
\seealso{
\code{\link{huge.NPN}} and \code{\link{huge}}
}