% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huge.roc.R
\name{print.roc}
\alias{print.roc}
\title{Print function for S3 class "roc"}
\usage{
\method{print}{roc}(x, ...)
}
\arguments{
\item{x}{An object with S3 class \code{"roc"}.}

\item{\dots}{System reserved (No specific usage)}
}
\description{
Print the information about true positive rates, false positive rates, the area under curve and maximum F1 score.
}
\seealso{
\code{\link{huge.roc}}
}
