% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huge.select.R
\name{print.select}
\alias{print.select}
\title{Print function for S3 class "select"}
\usage{
\method{print}{select}(x, ...)
}
\arguments{
\item{x}{An object with S3 class \code{"select"}.}

\item{\dots}{System reserved (No specific usage)}
}
\description{
Print the information about the model usage, graph dimension, model selection criterion, sparsity level of the optimal graph.
}
\seealso{
\code{\link{huge.select}}
}
