% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huge.generator.R
\name{print.sim}
\alias{print.sim}
\title{Print function for S3 class "sim"}
\usage{
\method{print}{sim}(x, ...)
}
\arguments{
\item{x}{An object with S3 class \code{"sim"}.}

\item{\dots}{System reserved (No specific usage)}
}
\description{
Print the information about the sample size, the dimension, the pattern and sparsity of the true graph structure.
}
\seealso{
\code{\link{huge.generator}}
}
