\name{summary.npn}
\alias{summary.npn}
\title{
Summary function for S3 class "npn"
}
\description{
Summary the information about the transformation function}
\usage{
\method{summary}{npn}(object, ...)
}
\arguments{
  \item{object}{
An object with S3 class \code{"npn"}
}
  \item{\dots}{
System reserved (No specific usage)
}
}
\author{
Tuo Zhao, Han Liu, Kathryn Roeder, John Lafferty, and Larry Wasserman \cr
Maintainers: Tuo Zhao<tourzhao@gmail.com>; Han Liu <hanliu@cs.jhu.edu>
}
\seealso{
\code{\link{huge.npn}}
}