\name{summary.sim}
\alias{summary.sim}
\title{
summary function for S3 class "sim"
}
\description{
summary the information about the sample size, the dimension, the pattern and sparsity of the true graph strcture.
}
\usage{
\method{summary}{sim}(object, ...)
}
\arguments{
  \item{object}{
An object with S3 class \code{"sim"}
}
  \item{\dots}{
System reserved (No specific usage)
}
}
\author{
Tuo Zhao, Han Liu, Kathryn Roeder, John Lafferty, and Larry Wasserman \cr
Maintainers: Tuo Zhao<tourzhao@andrew.cmu.edu>; Han Liu <hanliu@cs.jhu.edu>
}
\seealso{
\code{\link{huge.generator}} and  and \code{\link{huge}}
}